+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## **Release v1.58.0 - 2026-09-22**
 **Updated for Hearthstone 36.6.1**
 
 **Hearthstone**: